rm renderer
